Use new test syntax for space action mixins
