+++ packages/web-app-files/tests/unit/mixins/spaces/disable.spec.ts
@@ -1,16 +1,19 @@
-import Vuex from 'vuex'
-import { createStore } from 'vuex-extensions'
-import { mount, createLocalVue } from '@vue/test-utils'
 import disable from 'web-app-files/src/mixins/spaces/actions/disable.js'
-import { createLocationSpaces } from '../../../../src/router'
-import mockAxios from 'jest-mock-axios'
 import { buildSpace } from 'web-client/src/helpers'
-
-const localVue = createLocalVue()
-localVue.use(Vuex)
+import {
+  createStore,
+  defaultComponentMocks,
+  defaultPlugins,
+  mockAxiosResolve,
+  mount
+} from 'web-test-helpers'
+import { mockDeep } from 'jest-mock-extended'
+import { Graph } from 'web-client'
+import { clientService } from 'web-pkg'
+import { defaultStoreMockOptions } from 'web-test-helpers/src/mocks/store/defaultStoreMockOptions'
 
 const Component = {
-  render() {},
+  template: '<div></div>',
   mixins: [disable]
 }
 
@@ -19,7 +22,7 @@ describe('disable', () => {
 
   describe('isEnabled property', () => {
     it('should be false when no resource given', () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       expect(wrapper.vm.$_disable_items[0].isEnabled({ resources: [] })).toBe(false)
     })
     it('should be true when the space is not disabled', () => {
@@ -29,7 +32,7 @@ describe('disable', () => {
           permissions: [{ roles: ['manager'], grantedTo: [{ user: { id: 1 } }] }]
         }
       }
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       expect(wrapper.vm.$_disable_items[0].isEnabled({ resources: [buildSpace(spaceMock)] })).toBe(
         true
       )
@@ -42,7 +45,7 @@ describe('disable', () => {
           deleted: { state: 'trashed' }
         }
       }
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       expect(wrapper.vm.$_disable_items[0].isEnabled({ resources: [buildSpace(spaceMock)] })).toBe(
         false
       )
@@ -54,7 +57,7 @@ describe('disable', () => {
           permissions: [{ roles: ['viewer'], grantedTo: [{ user: { id: 1 } }] }]
         }
       }
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       expect(wrapper.vm.$_disable_items[0].isEnabled({ resources: [buildSpace(spaceMock)] })).toBe(
         false
       )
@@ -63,14 +66,14 @@ describe('disable', () => {
 
   describe('method "$_disable_trigger"', () => {
     it('should trigger the disable modal window', async () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const spyCreateModalStub = jest.spyOn(wrapper.vm, 'createModal')
       await wrapper.vm.$_disable_trigger({ resources: [{ id: 1 }] })
 
       expect(spyCreateModalStub).toHaveBeenCalledTimes(1)
     })
     it('should not trigger the disable modal window without any resource', async () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const spyCreateModalStub = jest.spyOn(wrapper.vm, 'createModal')
       await wrapper.vm.$_disable_trigger({ resources: [] })
 
@@ -80,11 +83,9 @@ describe('disable', () => {
 
   describe('method "$_disable_disableSpace"', () => {
     it('should hide the modal on success', async () => {
-      mockAxios.request.mockImplementationOnce(() => {
-        return Promise.resolve()
-      })
-
-      const wrapper = getWrapper()
+      const graphMock = mockDeep<Graph>()
+      graphMock.drives.deleteDrive.mockResolvedValue(mockAxiosResolve())
+      const { wrapper } = getWrapper(graphMock)
       const hideModalStub = jest.spyOn(wrapper.vm, 'hideModal')
       await wrapper.vm.$_disable_disableSpace(1)
 
@@ -92,12 +93,10 @@ describe('disable', () => {
     })
 
     it('should show message on error', async () => {
-      jest.spyOn(console, 'error').mockImplementation(() => {})
-      mockAxios.request.mockImplementationOnce(() => {
-        return Promise.reject(new Error())
-      })
-
-      const wrapper = getWrapper()
+      jest.spyOn(console, 'error').mockImplementation(() => undefined)
+      const graphMock = mockDeep<Graph>()
+      graphMock.drives.deleteDrive.mockRejectedValue(new Error())
+      const { wrapper } = getWrapper(graphMock)
       const showMessageStub = jest.spyOn(wrapper.vm, 'showMessage')
       await wrapper.vm.$_disable_disableSpace(1)
 
@@ -106,57 +105,19 @@ describe('disable', () => {
   })
 })
 
-function getWrapper() {
-  return mount(Component, {
-    localVue,
-    mocks: {
-      $router: {
-        currentRoute: createLocationSpaces('files-spaces-projects'),
-        resolve: (r) => {
-          return { href: r.name }
-        },
-        push: jest.fn()
-      },
-      $gettext: jest.fn(),
-      $gettextInterpolate: jest.fn()
-    },
-    store: createStore(Vuex.Store, {
-      actions: {
-        createModal: jest.fn(),
-        hideModal: jest.fn(),
-        showMessage: jest.fn(),
-        setModalInputErrorMessage: jest.fn()
-      },
-      getters: {
-        configuration: () => ({
-          server: 'https://example.com'
-        })
-      },
-      modules: {
-        user: {
-          state: {
-            id: 'alice',
-            uuid: 1
-          }
-        },
-        runtime: {
-          namespaced: true,
-          modules: {
-            auth: {
-              namespaced: true,
-              getters: {
-                accessToken: () => ''
-              }
-            },
-            spaces: {
-              namespaced: true,
-              mutations: {
-                UPDATE_SPACE_FIELD: jest.fn()
-              }
-            }
-          }
-        }
+function getWrapper(graphMock = mockDeep<Graph>()) {
+  jest.spyOn(clientService, 'graphAuthenticated').mockImplementation(() => graphMock)
+  const storeOptions = {
+    ...defaultStoreMockOptions,
+    modules: { ...defaultStoreMockOptions.modules, user: { state: { id: 'alice', uuid: 1 } } }
+  }
+  const store = createStore(storeOptions)
+  return {
+    wrapper: mount(Component, {
+      global: {
+        plugins: [...defaultPlugins(), store],
+        mocks: defaultComponentMocks({ currentRoute: { name: 'files-spaces-projects' } })
       }
     })
-  })
+  }
 }
