+++ packages/web-app-files/tests/unit/mixins/spaces/navigate.spec.ts
@@ -1,0 +1,72 @@
+import Navigate from 'web-app-files/src/mixins/spaces/actions/navigate.js'
+import { createLocationSpaces } from '../../../../src/router'
+import { createStore, defaultComponentMocks, defaultPlugins, mount } from 'web-test-helpers'
+import { defaultStoreMockOptions } from 'web-test-helpers/src/mocks/store/defaultStoreMockOptions'
+
+const Component = {
+  template: '<div></div>',
+  mixins: [Navigate]
+}
+
+describe('navigate', () => {
+  afterEach(() => jest.clearAllMocks())
+
+  describe('isEnabled property', () => {
+    it('should be true when no resource given', () => {
+      const { wrapper } = getWrapper()
+      expect(wrapper.vm.$_navigate_space_items[0].isEnabled({ resources: [] })).toBe(true)
+    })
+    it('should be false when resource is given', () => {
+      const { wrapper } = getWrapper()
+      expect(wrapper.vm.$_navigate_space_items[0].isEnabled({ resources: [{}] })).toBe(false)
+    })
+    it('should be false when location is invalid', () => {
+      const { wrapper } = getWrapper({ invalidLocation: true })
+      expect(wrapper.vm.$_navigate_space_items[0].isEnabled({ resources: [] })).toBe(false)
+    })
+  })
+
+  describe('method "$_navigate_space_trigger"', () => {
+    it('should trigger route change', async () => {
+      const { wrapper } = getWrapper()
+      const resource = { driveAlias: 'project/mars' }
+      await wrapper.vm.$_navigate_space_trigger({ resources: [resource] })
+
+      expect(wrapper.vm.$router.push).toHaveBeenCalledWith(
+        createLocationSpaces('files-spaces-generic', {
+          params: {
+            driveAliasAndItem: resource.driveAlias
+          },
+          query: {}
+        })
+      )
+    })
+  })
+})
+
+function getWrapper({ invalidLocation = false } = {}) {
+  const currentRoute = { name: '', params: { driveAliasAndItem: '' } }
+  if (invalidLocation) {
+    currentRoute.name = 'files-spaces-generic'
+  } else {
+    currentRoute.name = 'files-trash-generic'
+    currentRoute.params.driveAliasAndItem = 'project/mars'
+  }
+  const mocks = {
+    ...defaultComponentMocks({ currentRoute }),
+    space: {
+      driveType: 'project',
+      getDriveAliasAndItem: () => 'project/mars'
+    }
+  }
+  const storeOptions = defaultStoreMockOptions
+  const store = createStore(storeOptions)
+  return {
+    wrapper: mount(Component, {
+      global: {
+        plugins: [...defaultPlugins(), store],
+        mocks
+      }
+    })
+  }
+}
