+++ packages/web-app-files/tests/unit/mixins/spaces/rename.spec.ts
@@ -1,15 +1,18 @@
-import Vuex from 'vuex'
-import { createStore } from 'vuex-extensions'
-import { mount, createLocalVue } from '@vue/test-utils'
 import rename from 'web-app-files/src/mixins/spaces/actions/rename.js'
-import { createLocationSpaces } from '../../../../src/router'
-import mockAxios from 'jest-mock-axios'
-
-const localVue = createLocalVue()
-localVue.use(Vuex)
+import { mockDeep } from 'jest-mock-extended'
+import { Graph } from 'web-client'
+import { clientService } from 'web-pkg'
+import { defaultStoreMockOptions } from 'web-test-helpers/src/mocks/store/defaultStoreMockOptions'
+import {
+  createStore,
+  defaultComponentMocks,
+  defaultPlugins,
+  mockAxiosResolve,
+  mount
+} from 'web-test-helpers'
 
 const Component = {
-  render() {},
+  template: '<div></div>',
   mixins: [rename]
 }
 
@@ -18,14 +21,14 @@ describe('rename', () => {
 
   describe('method "$_rename_trigger"', () => {
     it('should trigger the rename modal window', async () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const spyCreateModalStub = jest.spyOn(wrapper.vm, 'createModal')
       await wrapper.vm.$_rename_trigger({ resources: [{ id: 1, name: 'renamed space' }] })
 
       expect(spyCreateModalStub).toHaveBeenCalledTimes(1)
     })
     it('should not trigger the rename modal window without any resource', async () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const spyCreateModalStub = jest.spyOn(wrapper.vm, 'createModal')
       await wrapper.vm.$_rename_trigger({ resources: [] })
 
@@ -35,14 +38,14 @@ describe('rename', () => {
 
   describe('method "$_rename_checkName"', () => {
     it('should throw an error with an empty space name', async () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const spyInputErrorMessageStub = jest.spyOn(wrapper.vm, 'setModalInputErrorMessage')
       await wrapper.vm.$_rename_checkName('')
 
       expect(spyInputErrorMessageStub).toHaveBeenCalledTimes(1)
     })
     it('should throw an error with an space name longer than 255 characters', async () => {
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const spyInputErrorMessageStub = jest.spyOn(wrapper.vm, 'setModalInputErrorMessage')
       await wrapper.vm.$_rename_checkName('n'.repeat(256))
 
@@ -51,7 +54,7 @@ describe('rename', () => {
     it.each(['/', '\\', '.', ':', '?', '*', '"', '>', '<', '|'])(
       'should show an error message when trying to create a space with a special character',
       (specialChar) => {
-        const wrapper = getWrapper()
+        const { wrapper } = getWrapper()
         wrapper.vm.setModalInputErrorMessage = jest.fn()
 
         const spyInputErrorMessageStub = jest.spyOn(wrapper.vm, 'setModalInputErrorMessage')
@@ -64,11 +67,9 @@ describe('rename', () => {
 
   describe('method "$_rename_renameSpace"', () => {
     it('should hide the modal and show message on success', async () => {
-      mockAxios.request.mockImplementationOnce(() => {
-        return Promise.resolve()
-      })
-
-      const wrapper = getWrapper()
+      const graphMock = mockDeep<Graph>()
+      graphMock.drives.updateDrive.mockResolvedValue(mockAxiosResolve())
+      const { wrapper } = getWrapper(graphMock)
       const hideModalStub = jest.spyOn(wrapper.vm, 'hideModal')
       const showMessageStub = jest.spyOn(wrapper.vm, 'showMessage')
       await wrapper.vm.$_rename_renameSpace(1, 'renamed space')
@@ -78,12 +79,10 @@ describe('rename', () => {
     })
 
     it('should show message on error', async () => {
-      jest.spyOn(console, 'error').mockImplementation(() => {})
-      mockAxios.request.mockImplementationOnce(() => {
-        return Promise.reject(new Error())
-      })
-
-      const wrapper = getWrapper()
+      jest.spyOn(console, 'error').mockImplementation(() => undefined)
+      const graphMock = mockDeep<Graph>()
+      graphMock.drives.updateDrive.mockRejectedValue(new Error())
+      const { wrapper } = getWrapper(graphMock)
       const showMessageStub = jest.spyOn(wrapper.vm, 'showMessage')
       await wrapper.vm.$_rename_renameSpace(1, 'renamed space')
 
@@ -92,55 +91,19 @@ describe('rename', () => {
   })
 })
 
-function getWrapper() {
-  return mount(Component, {
-    localVue,
-    mocks: {
-      $router: {
-        currentRoute: createLocationSpaces('files-spaces-projects'),
-        resolve: (r) => {
-          return { href: r.name }
-        }
-      },
-      $gettext: jest.fn()
-    },
-    store: createStore(Vuex.Store, {
-      actions: {
-        createModal: jest.fn(),
-        hideModal: jest.fn(),
-        showMessage: jest.fn(),
-        setModalInputErrorMessage: jest.fn()
-      },
-      getters: {
-        configuration: () => ({
-          server: 'https://example.com'
-        })
-      },
-      modules: {
-        Files: {
-          namespaced: true,
-          mutations: {
-            UPDATE_RESOURCE_FIELD: jest.fn()
-          }
-        },
-        runtime: {
-          namespaced: true,
-          modules: {
-            auth: {
-              namespaced: true,
-              getters: {
-                accessToken: () => ''
-              }
-            },
-            spaces: {
-              namespaced: true,
-              mutations: {
-                UPDATE_SPACE_FIELD: jest.fn()
-              }
-            }
-          }
-        }
+function getWrapper(graphMock = mockDeep<Graph>()) {
+  jest.spyOn(clientService, 'graphAuthenticated').mockImplementation(() => graphMock)
+  const storeOptions = {
+    ...defaultStoreMockOptions,
+    modules: { ...defaultStoreMockOptions.modules, user: { state: { id: 'alice', uuid: 1 } } }
+  }
+  const store = createStore(storeOptions)
+  return {
+    wrapper: mount(Component, {
+      global: {
+        plugins: [...defaultPlugins(), store],
+        mocks: defaultComponentMocks({ currentRoute: { name: 'files-spaces-projects' } })
       }
     })
-  })
+  }
 }
