+++ packages/web-app-files/tests/unit/mixins/spaces/uploadImage.spec.ts
@@ -1,6 +1,3 @@
-import Vuex from 'vuex'
-import { createStore } from 'vuex-extensions'
-import { mount, createLocalVue } from '@vue/test-utils'
 import uploadImage from 'web-app-files/src/mixins/spaces/actions/uploadImage.js'
 import { thumbnailService } from '../../../../src/services'
 import { mockDeep } from 'jest-mock-extended'
@@ -9,48 +6,14 @@ import { Graph } from 'web-client'
 import { clientService } from 'web-pkg'
 import { defaultComponentMocks } from 'web-test-helpers/src/mocks/defaultComponentMocks'
 import { defaultStoreMockOptions } from 'web-test-helpers/src/mocks/store/defaultStoreMockOptions'
+import { createStore, defaultPlugins, mount } from 'web-test-helpers'
 
-const localVue = createLocalVue()
-localVue.use(Vuex)
+const Component = {
+  template: '<div></div>',
+  mixins: [uploadImage]
+}
 
 describe('uploadImage', () => {
-  const Component = {
-    template: '<div></div>',
-    mixins: [uploadImage]
-  }
-
-  function getWrapper(resolvePutFileContents = true) {
-    const clientMock = mockDeep<OwnCloudSdk>()
-    const defaultMocks = defaultComponentMocks({ currentRoute: { name: 'files-spaces-generic' } })
-    return mount(Component, {
-      localVue,
-      data: () => ({
-        $_uploadImage_selectedSpace: {
-          id: '1fe58d8b-aa69-4c22-baf7-97dd57479f22'
-        }
-      }),
-      mocks: {
-        ...defaultMocks,
-        $client: {
-          ...clientMock,
-          files: {
-            ...clientMock.files,
-            putFileContents: jest.fn().mockImplementation(() => {
-              if (resolvePutFileContents) {
-                return Promise.resolve({
-                  'OC-FileId':
-                    'YTE0ODkwNGItNTZhNy00NTQ4LTk2N2MtZjcwZjhhYTY0Y2FjOmQ4YzMzMmRiLWUxNWUtNDRjMy05NGM2LTViYjQ2MGMwMWJhMw=='
-                })
-              }
-              return Promise.reject(new Error(''))
-            })
-          }
-        }
-      },
-      store: createStore(Vuex.Store, defaultStoreMockOptions)
-    })
-  }
-
   beforeAll(() => {
     thumbnailService.initialize({
       enabled: true,
@@ -69,7 +32,7 @@ describe('uploadImage', () => {
       })
       jest.spyOn(clientService, 'graphAuthenticated').mockImplementation(() => graphMock)
 
-      const wrapper = getWrapper()
+      const { wrapper } = getWrapper()
       const showMessageStub = jest.spyOn(wrapper.vm, 'showMessage')
       await wrapper.vm.$_uploadImage_uploadImageSpace({
         currentTarget: {
@@ -82,7 +45,7 @@ describe('uploadImage', () => {
 
     it('should show message on error', async () => {
       jest.spyOn(console, 'error').mockImplementation(() => undefined)
-      const wrapper = getWrapper(false)
+      const { wrapper } = getWrapper(false)
       const showMessageStub = jest.spyOn(wrapper.vm, 'showMessage')
       await wrapper.vm.$_uploadImage_uploadImageSpace({
         currentTarget: {
@@ -94,3 +57,41 @@ describe('uploadImage', () => {
     })
   })
 })
+
+function getWrapper(resolvePutFileContents = true) {
+  const clientMock = mockDeep<OwnCloudSdk>()
+  const defaultMocks = defaultComponentMocks({ currentRoute: { name: 'files-spaces-generic' } })
+  const mocks = {
+    ...defaultMocks,
+    $client: {
+      ...clientMock,
+      files: {
+        ...clientMock.files,
+        putFileContents: jest.fn().mockImplementation(() => {
+          if (resolvePutFileContents) {
+            return Promise.resolve({
+              'OC-FileId':
+                'YTE0ODkwNGItNTZhNy00NTQ4LTk2N2MtZjcwZjhhYTY0Y2FjOmQ4YzMzMmRiLWUxNWUtNDRjMy05NGM2LTViYjQ2MGMwMWJhMw=='
+            })
+          }
+          return Promise.reject(new Error(''))
+        })
+      }
+    }
+  }
+  const storeOptions = defaultStoreMockOptions
+  const store = createStore(storeOptions)
+  return {
+    wrapper: mount(Component, {
+      data: () => ({
+        $_uploadImage_selectedSpace: {
+          id: '1fe58d8b-aa69-4c22-baf7-97dd57479f22'
+        }
+      }),
+      global: {
+        plugins: [...defaultPlugins(), store],
+        mocks
+      }
+    })
+  }
+}
